+++ src/main/java/com/planz/planit/src/apicontroller/UserController.java
@@ -7,8 +7,8 @@ import com.planz.planit.src.domain.deviceToken.dto.DeviceTokenReqDTO;
 import com.planz.planit.src.domain.user.dto.*;
 import com.planz.planit.src.service.JwtTokenService;
 import com.planz.planit.src.service.UserService;
-import com.planz.planit.src.apicontroller.utils.ValidationRegex;
-import com.planz.planit.src.apicontroller.utils.ValidationUtils;
+import com.planz.planit.utils.ValidationRegex;
+import com.planz.planit.utils.ValidationUtils;
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
